moves spline helper logic into SplineComponent and deletes SplineHelperComponent
